Changed file structure of components, add css for login and signup, added logout to profile page
