+++ client/src/components/Login/Login.css
@@ -3,7 +3,7 @@
 }
 
 @media screen and (min-width: 768px) {
-        .signup {
-        padding-top: 60px;
+        .login {
+        padding-top: 100px;
     }
 }
