--- client/src/components/Profile.js
@@ -1,31 +1,0 @@
-import React, { Component } from 'react';
-import withAuth from './withAuth';
-import API from '../utils/API';
-import { Link } from 'react-router-dom';
-
-class Profile extends Component {
-
-  state = {
-    email: ""
-  };
-
-  componentDidMount() {
-    API.getUser(this.props.user.id).then(res => {
-      this.setState({
-        email: res.data.email
-      })
-    });
-  }
-
-  render() {
-    return (
-      <div className="container Profile">
-        <h1>On the profile page!</h1>
-        <p>Email: {this.state.email}</p>
-        <Link to="/">Go home</Link>
-      </div>
-    )
-  }
-}
-
-export default withAuth(Profile);
