+++ client/src/components/Profile/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Profile";
